requirements.txt
asgiref==3.8.1
boto3==1.35.36
botocore==1.35.36
certifi==2024.8.30
charset-normalizer==3.4.0
Django==5.1.1
django-cors-headers==4.4.0
django-environ==0.11.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
idna==3.10
jmespath==1.0.1
numpy==2.1.2
openpyxl==3.1.5
pandas==2.2.3
PyJWT==2.9.0
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
s3transfer==0.10.3
six==1.16.0
sqlparse==0.5.1
tzdata==2024.2
urllib3==2.2.3
